+++ lib/view/widgets/instruments_stats.dart
